Make library es6 friendly.
